requirements.txt
nicegui
python-sane # Also needs libsane-dev on Debian/Ubuntu
pillow
pypdf
requests